Migrate translations clean script (#33930)
